Made changes to Entity class - Now has an attribute sprites to know how many sprites to animate Added Repeater class

+++ Shooter.cpp
@@ -1,6 +1,6 @@
 #include "Shooter.h"
 
-Shooter::Shooter(int x, int y, int reloadTime, int bulletSpeed, string bulletSpritePath, int textureX, int textureY) :pos(x, y), reloadTime(reloadTime), bulletSpeed(bulletSpeed), bulletSpritePath(bulletSpritePath),textureX(textureX),textureY(textureY) {
+Shooter::Shooter(int x, int y, float reloadTime, int bulletSpeed, string bulletSpritePath, int textureX, int textureY) :pos(x, y), reloadTime(reloadTime), bulletSpeed(bulletSpeed), bulletSpritePath(bulletSpritePath),textureX(textureX),textureY(textureY) {
 	bullet = nullptr;
 }
