+++ Shooter.h
@@ -8,13 +8,13 @@ using namespace std;
 class Shooter {
 protected:
 	Bullet* bullet;
-	int reloadTime,  bulletSpeed,textureX,textureY;
+	float reloadTime,  bulletSpeed,textureX,textureY;
 	string bulletSpritePath;
 	Clock clock;
 	Position pos;
 
 public:
-	Shooter(int x,int y, int reloadTime, int bulletSpeed, string bulletSpritePath, int textureX, int textureY);
+	Shooter(int x,int y, float reloadTime, int bulletSpeed, string bulletSpritePath, int textureX, int textureY);
 
 	Bullet* shoot();
